pyproject.toml
[project]
name = "enhanced-qna-agent"
version = "1.0.0"
description = "Enhanced QnA Agent System - MVP"
requires-python = ">=3.9"

dependencies = [
    # Core dependencies
    "streamlit>=1.28.0",
    "flask>=2.3.0",
    "flask-socketio>=5.3.0",
    "flask-cors>=4.0.0",
    "python-socketio[client]>=5.8.0",
    "python-engineio>=4.7.0",
    # Database
    "psycopg2-binary>=2.9.0",
    # Data processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    # AI/ML
    "langchain>=0.0.300",
    "langchain-openai>=0.0.2",
    "openai>=1.0.0",
    # HTTP requests
    "requests>=2.31.0",
    # Environment and configuration
    "python-dotenv>=1.0.0",
    # Utilities
    "python-dateutil>=2.8.0",
    "websocket-client>=1.8.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
start-backend = "scripts.start_backend:main"
start-frontend = "scripts.start_frontend:main"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
